third_party/WebKit/LayoutTests/media/media-controller-playbackrate-expected.txt

EVENT(canplay)
TEST(video.controller.defaultPlaybackRate = -Infinity) THROWS("TypeError: Failed to set the 'defaultPlaybackRate' property on 'MediaController': The provided double value is non-finite.") OK
TEST(video.controller.defaultPlaybackRate = Infinity) THROWS("TypeError: Failed to set the 'defaultPlaybackRate' property on 'MediaController': The provided double value is non-finite.") OK
TEST(video.controller.defaultPlaybackRate = NaN) THROWS("TypeError: Failed to set the 'defaultPlaybackRate' property on 'MediaController': The provided double value is non-finite.") OK
TEST(video.controller.playbackRate = -Infinity) THROWS("TypeError: Failed to set the 'playbackRate' property on 'MediaController': The provided double value is non-finite.") OK
TEST(video.controller.playbackRate = Infinity) THROWS("TypeError: Failed to set the 'playbackRate' property on 'MediaController': The provided double value is non-finite.") OK
TEST(video.controller.playbackRate = NaN) THROWS("TypeError: Failed to set the 'playbackRate' property on 'MediaController': The provided double value is non-finite.") OK
EVENT(seeked)
EVENT(play)
EVENT(timeupdate)
EXPECTED (video.currentTime != '0') OK
END OF TEST
